move config initialization to config module
