+++ frontend/src/components/admin/configuration/AdminConfigInput.tsx
@@ -44,6 +44,7 @@ const AdminConfigInput = ({
             style={{
               width: "100%",
             }}
+            disabled={!configVariable.allowEdit}
             {...form.getInputProps("stringValue")}
             onChange={(e) => onValueChange(configVariable, e.target.value)}
           />
@@ -52,6 +53,7 @@ const AdminConfigInput = ({
             style={{
               width: "100%",
             }}
+            disabled={!configVariable.allowEdit}
             {...form.getInputProps("stringValue")}
             placeholder={configVariable.defaultValue}
             onChange={(e) => onValueChange(configVariable, e.target.value)}
@@ -63,6 +65,7 @@ const AdminConfigInput = ({
           style={{
             width: "100%",
           }}
+          disabled={!configVariable.allowEdit}
           autosize
           {...form.getInputProps("textValue")}
           placeholder={configVariable.defaultValue}
@@ -72,6 +75,7 @@ const AdminConfigInput = ({
       {configVariable.type == "number" && (
         <NumberInput
           {...form.getInputProps("numberValue")}
+          disabled={!configVariable.allowEdit}
           placeholder={configVariable.defaultValue}
           onChange={(number) => onValueChange(configVariable, number)}
           w={201}
@@ -80,6 +84,7 @@ const AdminConfigInput = ({
       {configVariable.type == "filesize" && (
         <FileSizeInput
           {...form.getInputProps("numberValue")}
+          disabled={!configVariable.allowEdit}
           value={parseInt(configVariable.value ?? configVariable.defaultValue)}
           onChange={(bytes) => onValueChange(configVariable, bytes)}
           w={201}
@@ -88,6 +93,7 @@ const AdminConfigInput = ({
       {configVariable.type == "boolean" && (
         <>
           <Switch
+            disabled={!configVariable.allowEdit}
             {...form.getInputProps("booleanValue", { type: "checkbox" })}
             onChange={(e) => onValueChange(configVariable, e.target.checked)}
           />
@@ -96,6 +102,7 @@ const AdminConfigInput = ({
       {configVariable.type == "timespan" && (
         <TimespanInput
           value={stringToTimespan(configVariable.value)}
+          disabled={!configVariable.allowEdit}
           onChange={(timespan) =>
             onValueChange(configVariable, timespanToString(timespan))
           }
